refactor: extract MenuProductInfoView from MenuBarView
Move the Product Info / Help & Support section (and the option-key
debug-bundle affordance) out of the 1,066-line MenuBarView into a focused
MenuProductInfoView under Sources/BugNarrator/Views/Menu/. Behavior-
preserving: the section renders identically and every action still
delegates to AppState.

First section split for #433 (#457).
